requirements.txt
3==0.0.0
altgraph==0.17
backports.entry-points-selectable==1.1.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cycler==0.11.0
distlib==0.3.2
docutils==0.17.1
filelock==3.0.12
future==0.18.2
idna==2.10
importlib-metadata==3.7.0
importlib-resources==5.2.2
keyboard==0.13.5
Kivy==2.0.0
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.sdl2==0.3.1
Kivy-examples==2.0.0
Kivy-Garden==0.1.4
kiwisolver==1.3.1
matplotlib==3.3.4
MouseInfo==0.1.3
mss==6.1.0
numpy==1.19.5
opencv-python==4.5.1.48
pefile==2019.4.18
Pillow==8.2.0
platformdirs==2.2.0
PyAutoGUI==0.9.52
pycparser==2.20
PyGetWindow==0.0.9
Pygments==2.9.0
PyInstaller==3.6
pyinstaller-hooks-contrib==2020.11
PyMsgBox==1.0.9
pyparsing==3.0.6
pyperclip==1.8.1
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.26
pyserial==3.5
pyTelegramBotAPI==3.7.6
python-dateutil==2.8.2
PyTweening==1.0.3
pyusb==1.1.1
pywin32==301
pywin32-ctypes==0.2.0
requests==2.25.1
six==1.16.0
sounddevice==0.4.1
style==1.1.0
typing-extensions==3.7.4.3
update==0.0.1
urllib3==1.26.3
virtualenv==20.7.2
WinUsbCDC==1.4
zipp==3.4.0

Django~=3.2.12
djangorestframework~=3.13.1